Update api proxy classes to typescript and use mojom types directly in them
