+++ components/brave_wallet_ui/page/BUILD.gn
@@ -19,7 +19,6 @@ transpile_web_ui("brave_wallet_page_ui") {
   resource_name = "brave_wallet_page"
   deps = [
     ":preprocess_page",
-    ":wallet_page_api_proxy",
     "//brave/components/brave_wallet/common:mojom_js",
     "//brave/components/brave_wallet/common:preprocess_mojo",
   ]
@@ -35,13 +34,5 @@ preprocess_if_expr("preprocess_page") {
   in_folder = "./"
   out_folder = "$target_gen_dir/$preprocess_folder"
   out_manifest = "$target_gen_dir/$preprocess_page_manifest"
-  in_files = [ "wallet_page_api_proxy.js" ]
-}
-
-js_library("wallet_page_api_proxy") {
-  deps = [
-    "//brave/components/brave_wallet/common:mojom_js_library_for_compile",
-    "//ui/webui/resources/js:cr.m",
-    "//url/mojom:url_mojom_gurl_js_library_for_compile",
-  ]
+  in_files = [ "wallet_page_api_proxy.ts" ]
 }
