+++ components/brave_wallet_ui/panel/wallet_panel_api_proxy.ts
@@ -1,0 +1,53 @@
+// Copyright (c) 2021 The Brave Authors. All rights reserved.
+// This Source Code Form is subject to the terms of the Mozilla Public
+// License, v. 2.0. If a copy of the MPL was not distributed with this file,
+// you can obtain one at http://mozilla.org/MPL/2.0/.
+
+import WalletApiProxy, * as BraveWallet from '../common/wallet_api_proxy'
+
+class WalletPanelApiProxy extends WalletApiProxy {
+  callbackRouter = new BraveWallet.PageCallbackRouter()
+  panelHandler = new BraveWallet.PanelHandlerRemote()
+
+  constructor() {
+    super()
+
+    const factory = BraveWallet.PanelHandlerFactory.getRemote()
+    factory.createPanelHandler(
+        this.callbackRouter.$.bindNewPipeAndPassRemote(),
+        this.panelHandler.$.bindNewPipeAndPassReceiver(),
+        this.walletHandler.$.bindNewPipeAndPassReceiver(),
+        this.ethJsonRpcController.$.bindNewPipeAndPassReceiver(),
+        this.swapController.$.bindNewPipeAndPassReceiver(),
+        this.assetRatioController.$.bindNewPipeAndPassReceiver(),
+        this.keyringController.$.bindNewPipeAndPassReceiver(),
+        this.ercTokenRegistry.$.bindNewPipeAndPassReceiver(),
+        this.ethTxController.$.bindNewPipeAndPassReceiver(),
+        this.braveWalletService.$.bindNewPipeAndPassReceiver())
+  }
+
+  showUI() {
+    this.panelHandler.showUI()
+  }
+
+  closeUI() {
+    this.panelHandler.closeUI()
+  }
+
+  connectToSite(accounts: string[], origin: string, tab_id: number) {
+    this.panelHandler.connectToSite(accounts, origin, tab_id)
+  }
+
+  cancelConnectToSite(origin: string, tab_id: number) {
+    this.panelHandler.cancelConnectToSite(origin, tab_id)
+  }
+}
+
+let walletPanelApiProxyInstance: WalletPanelApiProxy
+
+export default function getWalletPanelApiProxy () {
+  if (!walletPanelApiProxyInstance) {
+    walletPanelApiProxyInstance = new WalletPanelApiProxy()
+  }
+  return walletPanelApiProxyInstance
+}
